better logout handling
see #360

+++ app/app/src/main/java/com/github/livingwithhippos/unchained/settings/viewmodel/SettingsViewModel.kt
@@ -28,7 +28,7 @@ constructor(
     private val pluginRepository: PluginRepository,
     private val kodiRepository: KodiRepository,
     private val protoStore: ProtoStore,
-    private val preferences: SharedPreferences
+    private val preferences: SharedPreferences,
 ) : ViewModel() {
 
     val kodiLiveData = MutableLiveData<Event<Boolean>>()
@@ -67,8 +67,13 @@ constructor(
 
     fun userLogout() {
         viewModelScope.launch {
-            protoStore.deleteCredentials()
-            eventLiveData.postEvent(SettingEvent.Logout)
+            val credentials = protoStore.getCredentials()
+            if (credentials.accessToken.isBlank() && credentials.clientId.isBlank()) {
+                eventLiveData.postEvent(SettingEvent.LogoutNoCredentials)
+            } else {
+                protoStore.deleteCredentials()
+                eventLiveData.postEvent(SettingEvent.Logout)
+            }
         }
     }
 
@@ -89,4 +94,6 @@ constructor(
 
 sealed class SettingEvent {
     data object Logout : SettingEvent()
+
+    data object LogoutNoCredentials : SettingEvent()
 }
